change lat/long to f32

+++ src/record.rs
@@ -4,9 +4,9 @@ pub struct Record {
     /// location name
     pub name: String,
     /// latitude
-    pub latitude: f64,
+    pub latitude: f32,
     /// longitude
-    pub longitude: f64,
+    pub longitude: f32,
     #[serde(rename="country code")]
     /// country code from geonamedb
     pub country: String,
